requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.1.1
pytest-html==3.1.1
pytest-metadata==2.0.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.7.0
